reorganize spec helper
